consolidated theme code into one provider
