+++ frontend/src/pages/console.tsx
@@ -1229,7 +1229,7 @@ const Console = () => {
 /**
  * Default component
  */
-/*
+
 export default function Page() {
   const [timezone, setTimezone] = useState('utc');
 
@@ -1240,26 +1240,6 @@ export default function Page() {
           <Console />
         </AppLayout>
       </Context.Provider>
-    </AuthRequired>
-  );
-}
-*/
-
-export default function Page() {
-  const { theme, userPreference, setUserPreference } = useTheme();
-
-  const handleChange = (ev: React.ChangeEvent<HTMLSelectElement>) => {
-    setUserPreference(ev.target.value as UserPreference);
-  };
-
-  return (
-    <AuthRequired>
-      <div>current theme: {theme}</div>
-      <select value={userPreference} onChange={handleChange}>
-        <option value={UserPreference.System}>system</option>
-        <option value={UserPreference.Dark}>dark</option>
-        <option value={UserPreference.Light}>light</option>
-      </select>
     </AuthRequired>
   );
 }
